6078: TOGGLE refaktor og legge til kobling mellom faktura og fakturamottatt (#89)
* Refaktorerer faktura mottatt

* Disabler test

* Refaktor litt

+++ src/main/kotlin/no/nav/faktureringskomponenten/service/mappers/FakturaMottattMapper.kt
@@ -11,7 +11,7 @@ class FakturaMottattMapper {
         fakturaMottatt: FakturaMottattDto
     ): FakturaMottatt = FakturaMottatt(
             id = null,
-            fakturaReferanseNr = fakturaMottatt.fakturaReferanseNr,
+            fakturaReferanseNr = fakturaMottatt.fakturaReferanseNr.toLong(),
             fakturaNummer = fakturaMottatt.fakturaNummer,
             dato = fakturaMottatt.dato,
             status = fakturaMottatt.status,
